validaciones crud marcas y categorias al editar

+++ www/controller/edit_category.php
@@ -5,6 +5,15 @@ $id = $_POST['edit-id'];
 $detail = $_POST['edit-detail'];
 $detail=strtoupper($detail);
 $status = 1;
+if(category_exists($detail))
+{
+  echo '<script>
+  localStorage.setItem("mensaje", "La categoria ya existe. Por favor, elija un nombre diferente");
+  localStorage.setItem("tipo", "error");
+  window.location.href = "../views/crud_category.php";
+      </script>';  
+}
+else {
  if (!Updatecategory($id, $detail, $status)) {
    echo '<script>
    localStorage.setItem("mensaje", "Categoría editada con éxito");
@@ -17,5 +26,6 @@ $status = 1;
    localStorage.setItem("tipo", "error");
    window.location.href = "../views/crud_category.php";
        </script>';   
+}
 }
 ?>
